pretty-hydra: better support for byte compilation
Previously if a hydra is modified using `pretty-hydra-define+` in the
same `.el` file as the original definition, byte compiling the .el
file would fail. That's because the auto-generated var
`hydra-name/heads-plist` only gets created at load time, but it's read
in `pretty-hydra-define+` at compile/macro expansion time.

Now the two defvar/set calls are wrapped in `eval-and-compile` so that
those vars are available at both compile time and load time.

Also always use `defhydra` as `defhydra+` has a similar issue. We
don't really need to use it for `pretty-hydra-define+` anyway, as all
the heads plist and body are already stored.