Use `plist-get` and `plist-put` with `#'equal` from Emacs/Compat 29
From Emacs 29 (currently in pretest), `plist-get` and `plist-put` take an
optional comparison predicate which we set to `#'equal`. Previously,
`lax-plist-get`/`-put` was used for that, but those functions are now marked as
obsolete.

To keep compatibility with older Emacs releases, I added a dependency on
Compat (https://elpa.gnu.org/packages/compat.html).